+++ test/end2end/pub_server_data.dart
@@ -80,7 +80,8 @@ final _data = {
         "package:pub_semver/src/version_union.dart"
       ],
       'platform': {
-        'worksEverywhere': true,
+        'components': [],
+        'uses': {'flutter': 'allowed', 'server': 'allowed', 'web': 'allowed'},
       },
       'fitness': {'magnitude': 95.0, 'shortcoming': 0.0},
     },
@@ -284,13 +285,15 @@ final _data = {
         "package:yaml/yaml.dart"
       ],
       'platform': {
-        'worksEverywhere': true,
+        'components': [],
+        'uses': {'flutter': 'allowed', 'server': 'allowed', 'web': 'allowed'},
       },
       'fitness': {'magnitude': 360.0, 'shortcoming': 1.0},
     },
   },
   'platform': {
-    'worksEverywhere': true,
+    'components': [],
+    'uses': {'flutter': 'allowed', 'server': 'allowed', 'web': 'allowed'},
     'reason': 'No platform restriction found in libraries.',
   },
   "licenses": [